sandbox/hiring-agent/backend/requirements.txt
streamlit
openai
firecrawl
PyPDF2
python-dotenv
pydantic
fastapi
uvicorn